apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "atgo-api"
version = "0.1.0"
description = "ATGO Cloud Attendance API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "pydantic>=2.10",
    "pydantic-settings>=2.6",
    "email-validator>=2.2",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "python-jose[cryptography]>=3.3",
    "python-multipart>=0.0.18",
    "redis>=5.2",
    "httpx>=0.28",
    "structlog>=24.4",
    "python-dateutil>=2.9",
    "openpyxl>=3.1",
    "dnspython>=2.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
]

[tool.setuptools.packages.find]
include = ["atgo_api*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
